requirements.txt
flask==1.0.2
gunicorn==19.9.0
PyYAML==5.1